Create file-management.asm
This code sets up a basic file system management system that uses a pointer to keep track of the next available memory location for storing new files, and functions to create, open and delete files. The code uses a simple linear search algorithm to find the desired file in the file system.